Bernabe's Initial Commit.
