+++ bin/Debug/netcoreapp3.0/RDLApp.runtimeconfig.dev.json
@@ -1,0 +1,8 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\Buildsmart IT\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\Buildsmart IT\\.nuget\\packages"
+    ]
+  }
+}
